Author: Bleep <tomh@inxpress.net> Log message: Audit usage of size_t values, clean up stuff for 64 bit. Fixup undefined ops.
git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/branches/u2_10_10_beta06@33 c9e4aea6-c8fd-4c43-8297-357d70d61c8c

+++ include/res.h
@@ -39,13 +39,13 @@ extern time_t          timeout_resolver(time_t now);
  * pointer arg, DO NOT call this from a resolver callback function the
  * resolver will delete the query itself for the affected client.
  */
-extern void            delete_resolver_queries(const void* vptr);
-extern unsigned long   cres_mem(struct Client* cptr);
-extern int             m_dns(struct Client* cptr, struct Client* sptr,
+extern void     delete_resolver_queries(const void* vptr);
+extern size_t   cres_mem(struct Client* cptr);
+extern int      m_dns(struct Client* cptr, struct Client* sptr,
                              int parc, char* parv[]);
-extern int             resolver_read(void);
-extern void            resolver_read_multiple(int count);
-extern void            flush_resolver_cache(void);
+extern int      resolver_read(void);
+extern void     resolver_read_multiple(int count);
+extern void     flush_resolver_cache(void);
 
 /*
  * add_local_domain - append local domain suffix to hostnames that
